move address table lookup compile + codec into transaction-messages
