+++ packages/transaction-messages/package.json
@@ -64,6 +64,9 @@
     ],
     "dependencies": {
         "@solana/addresses": "workspace:*",
+        "@solana/codecs-core": "workspace:*",
+        "@solana/codecs-data-structures": "workspace:*",
+        "@solana/codecs-numbers": "workspace:*",
         "@solana/errors": "workspace:*",
         "@solana/instructions": "workspace:*",
         "@solana/rpc-types": "workspace:*"
