+++ packages/transaction-messages/src/codecs/__tests__/address-table-lookup-test.ts
@@ -1,6 +1,6 @@
 import type { Address } from '@solana/addresses';
 
-import type { getCompiledAddressTableLookups } from '../../compile-address-table-lookups';
+import type { getCompiledAddressTableLookups } from '../../compile/address-table-lookups';
 import {
     getAddressTableLookupCodec,
     getAddressTableLookupDecoder,
