+++ packages/transaction-messages/src/codecs/index.ts
@@ -1,0 +1,1 @@
+export * from './address-table-lookup';
