+++ packages/transaction-messages/src/compile/__tests__/address-table-lookups-test.ts
@@ -1,8 +1,8 @@
 import { Address, getAddressComparator } from '@solana/addresses';
 import { AccountRole } from '@solana/instructions';
-import { OrderedAccounts } from '@solana/transaction-messages';
 
-import { getCompiledAddressTableLookups } from '../compile-address-table-lookups';
+import { getCompiledAddressTableLookups } from '../../compile/address-table-lookups';
+import { OrderedAccounts } from '../accounts';
 
 const MOCK_ADDRESSES: ReadonlyArray<Address> = [
     'BRwZRKsvKkG45g59269qZ5e8UaECFim5Qfxex44UKwDG',
