+++ packages/transaction-messages/src/compile/index.ts
@@ -1,1 +1,2 @@
 export * from './accounts';
+export * from './address-table-lookups';
