+++ packages/transactions/src/__tests__/message-test.ts
@@ -1,7 +1,7 @@
 import { Address } from '@solana/addresses';
+import { getCompiledAddressTableLookups } from '@solana/transaction-messages';
 
 import { ITransactionWithBlockhashLifetime } from '../blockhash';
-import { getCompiledAddressTableLookups } from '../compile-address-table-lookups';
 import { getCompiledMessageHeader } from '../compile-header';
 import { getCompiledInstructions } from '../compile-instructions';
 import { getCompiledLifetimeToken } from '../compile-lifetime-token';
@@ -10,7 +10,10 @@ import { ITransactionWithFeePayer } from '../fee-payer';
 import { compileTransactionMessage } from '../message';
 import { BaseTransaction } from '../types';
 
-jest.mock('../compile-address-table-lookups');
+jest.mock('@solana/transaction-messages', () => ({
+    ...jest.requireActual('@solana/transaction-messages'),
+    getCompiledAddressTableLookups: jest.fn(),
+}));
 jest.mock('../compile-header');
 jest.mock('../compile-instructions');
 jest.mock('../compile-lifetime-token');
@@ -76,6 +79,7 @@ describe('compileTransactionMessage', () => {
         });
         it('sets `instructions` to the return value of `getCompiledInstructions`', () => {
             const message = compileTransactionMessage(baseTx);
+            console.log({ message });
             expect(getCompiledInstructions).toHaveBeenCalledWith(
                 baseTx.instructions,
                 expect.any(Array) /* orderedAccounts */,
